// src/types/consent.ts
export interface IConsentContent {
  label: string;
  privacyPolicyLabel: string;
  privacyPolicyHref: string;
  disclaimer: string;
  readMoreLabel: string;
  readMoreIntro: string;
  readMoreItems: string[];
  readMoreFrequency: string;
  readMoreRates: string;
  readMoreOptOut: string;
  readMoreHelp: string;
  readMorePrivacyText: string;
}